requirements.txt
python-telegram-bot==21.3
beautifulsoup4
requests
